requirements.txt
Flask
snowflake-connector-python
bcrypt
openai
python-dotenv